Add highlighted dungeon management and refresh home highlights
