+++ nwleaderboard-ui/js/locales/en.js
@@ -34,6 +34,15 @@ const en = {
   contributeResultsTitle: 'Detected runs',
   contributeToggleSuccessLabel: 'Show validated items',
   contributeReadyForSubmission: 'Ready to submit.',
+  contributeMenuLabel: 'Contributor sections',
+  contributeMenuDungeons: 'Dungeons',
+  contributeMenuImport: 'Import',
+  contributeDungeonsDescription:
+    'Choose which dungeons are highlighted on the home page. Only selected dungeons will appear in the highlights.',
+  contributeHighlightsSave: 'Save selection',
+  contributeHighlightsSaving: 'Saving…',
+  contributeHighlightsSaved: 'Highlighted dungeons updated.',
+  contributeHighlightsError: 'Unable to update highlighted dungeons.',
   contributeRunLabel: (index) => `Run ${index}`,
   contributeExtractionReady: 'Runs extracted. Review them below before saving.',
   contributeNoResults: 'No run detected on this screenshot.',
@@ -130,6 +139,14 @@ const en = {
   versionError: 'Unable to verify the application version. Offline mode is active.',
   leaderboardTitle: 'Latest highlights',
   highlights: 'Highlights',
+  highlightLoading: 'Loading highlights…',
+  highlightError: 'Unable to load highlights.',
+  highlightEmpty: 'No highlighted dungeons yet. Visit the contributor area to select them.',
+  highlightScoreLabel: 'Best score',
+  highlightTimeLabel: 'Best time',
+  highlightNoScore: 'No score yet',
+  highlightNoTime: 'No time yet',
+  highlightWeekLabel: 'Week',
   leaderboardMenuTitle: 'Leaderboards',
   leaderboardEmpty: 'No scores available yet. Stay tuned!',
   navMenu: 'Menu',
