+++ nwleaderboard-ui/js/locales/fr.js
@@ -35,6 +35,15 @@ const fr = {
   contributeResultsTitle: 'Runs détectés',
   contributeToggleSuccessLabel: 'Afficher les éléments validés',
   contributeReadyForSubmission: 'Prêt pour envoi.',
+  contributeMenuLabel: 'Sections contributeur',
+  contributeMenuDungeons: 'Donjons',
+  contributeMenuImport: 'Importer',
+  contributeDungeonsDescription:
+    'Sélectionnez les donjons mis en avant sur la page d’accueil. Seuls les donjons cochés apparaîtront dans les faits marquants.',
+  contributeHighlightsSave: 'Enregistrer la sélection',
+  contributeHighlightsSaving: 'Enregistrement…',
+  contributeHighlightsSaved: 'Faits marquants mis à jour.',
+  contributeHighlightsError: 'Impossible de mettre à jour les faits marquants.',
   contributeRunLabel: (index) => `Run n°${index}`,
   contributeExtractionReady: 'Runs extraits. Vérifiez-les avant de valider.',
   contributeNoResults: 'Aucun run détecté sur cette capture.',
@@ -133,6 +142,15 @@ const fr = {
     "Impossible de vérifier la version de l'application. Le mode hors-ligne est actif.",
   leaderboardTitle: 'Derniers faits marquants',
   highlights: 'Faits marquants',
+  highlightLoading: 'Chargement des faits marquants…',
+  highlightError: 'Impossible de charger les faits marquants.',
+  highlightEmpty:
+    'Aucun donjon mis en avant pour le moment. Utilisez la zone contributeur pour les sélectionner.',
+  highlightScoreLabel: 'Meilleur score',
+  highlightTimeLabel: 'Meilleur temps',
+  highlightNoScore: 'Aucun score pour le moment',
+  highlightNoTime: 'Aucun temps pour le moment',
+  highlightWeekLabel: 'Semaine',
   leaderboardMenuTitle: 'Classements',
   leaderboardEmpty: 'Aucun score pour le moment. Revenez vite !',
   navMenu: 'Menu',
